[libc][posix/io/stdio] rename libc.c as posix/stdio.c
