+++ components/libc/compilers/newlib/syscalls.c
@@ -21,7 +21,7 @@
 #include <sys/errno.h>
 #include <sys/stat.h>
 #ifdef RT_USING_POSIX_STDIO
-#include "libc.h"
+#include <posix/stdio.h>
 #endif /* RT_USING_POSIX_STDIO */
 #ifdef RT_USING_MODULE
 #include <dlmodule.h>
@@ -225,7 +225,7 @@ _ssize_t _read_r(struct _reent *ptr, int fd, void *buf, size_t nbytes)
     if (fd == STDIN_FILENO)
     {
 #ifdef RT_USING_POSIX_STDIO
-        if (libc_stdio_get_console() < 0)
+        if (rt_posix_stdio_get_console() < 0)
         {
             LOG_W("Do not invoke standard input before initializing Compiler");
             return 0;
